requirements.txt
google-generativeai>=0.3.0
python-docx>=0.8.11 